test: 삭제 및 분리
