--- e2e/apiUrl.ts
@@ -1,1 +1,0 @@
-export const apiUrl = "https://www.pre-onboarding-selection-task.shop" as const
